+++ src/models/receipt_model.py
@@ -1,101 +1,8 @@
 from .abstract_reference import *
 from .range_model import range_model
 from .range_group_model import range_group_model
-
-
-class step():
-    """
-    Class that contain data of step of receipt's instruction 
-
-    step_description:str - info about step
-    """
-    __step_description:str
-
-    def __init__(self,step_description:str=None):
-        """
-        Constructor of class
-        step_description:str
-        """
-        self.step_description=step_description
-
-    @property
-    def step_description(self):
-        """
-        Function that returns step_description
-        """
-        return self.__step_description
-    
-    @step_description.setter
-    def step_description(self,value:str):
-        """
-        Function that set step_description
-        value:str
-        """
-        if model_validator.check_type(value,str):
-            self.__step_description=value
-        else:
-            raise argument_exception("step_description can be only string")
-
-    def __str__(self):
-        """
-        Magic function that returns string value of class object
-        """
-        return str(self.step_description)
-
-class proportion():
-    """
-    Class that contains info about proportion
-    range:range_model - range of proportion
-    proportion_value:float - value of range proportion  
-    """
-    __range:range_model
-    __proportion_value:float
-
-    def __init__(self,range:range_model=None,proportion_value:float=1.0):
-        """
-        Constructor of class
-        range:range_model
-        proportion_value:float
-        """
-        self.range=range
-        self.proportion_value=proportion_value
-
-
-    @property
-    def range(self):
-        """
-        Function that returns range
-        """
-        return self.__range
-    
-    @range.setter
-    def range(self,value:range_model):
-        """
-        Function that sets property range
-        value:range_model
-        """
-        if model_validator.check_type(value,range_model):
-            self.__range=value
-        else:
-            raise argument_exception("range should be range_model type")
-        
-    @property
-    def proportion_value(self):
-        """
-        Function that returns proportion_value
-        """
-        return self.__proportion_value
-    
-    @proportion_value.setter
-    def proportion_value(self,value:float):
-        """
-        Function that sets property proportion_value
-        value:float
-        """
-        if model_validator.check_type(value,float):
-            self.__proportion_value=value
-        else:
-            raise argument_exception("proportion_value should be float type")
+from .step import step
+from.proportion import proportion
 
 class receipt_model(abstract_reference):
     """
